+++ source/WindowBaseClass.cpp
@@ -19,7 +19,7 @@ void WindowBaseClass::Render()
 
 }
 
-void WindowBaseClass::Update(double p_deltaTime, float p_mousePositionX, float p_mousePositionY)
+void WindowBaseClass::Update(double p_deltaTime, int p_mousePositionX, int p_mousePositionY)
 {
 	m_gameInterface->Update(p_deltaTime, p_mousePositionX, p_mousePositionY);
 }
